Absorb label audit into backporter.jl
- Add --audit and --cleanup-pr flags to backporter.jl
- Auto-detect repo from git remote (no longer required)
- Use local git commands instead of GitHub API for fetching commits
- Remove PR comment functionality when removing labels
