+++ backporter.jl
@@ -60,43 +60,54 @@ struct CLIOptions
     validate_branch::Bool
     require_clean::Bool
     test_commit::Union{String, Nothing}
+    audit::Bool                         # Run label audit mode
+    cleanup_pr::Union{Int, Nothing}     # PR number for cleanup mode
 end
 
 function parse_cli_args(args::Vector{String})
-    options = CLIOptions(nothing, nothing, false, false, true, true, nothing)
+    options = CLIOptions(nothing, nothing, false, false, true, true, nothing, false, nothing)
 
     i = 1
     while i <= length(args)
         arg = args[i]
         if arg == "--help" || arg == "-h"
-            options = CLIOptions(options.version, options.repo, true, options.dry_run, options.validate_branch, options.require_clean, options.test_commit)
+            options = CLIOptions(options.version, options.repo, true, options.dry_run, options.validate_branch, options.require_clean, options.test_commit, options.audit, options.cleanup_pr)
         elseif arg == "--version" || arg == "-v"
             if i + 1 <= length(args)
-                options = CLIOptions(args[i+1], options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit)
+                options = CLIOptions(args[i+1], options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit, options.audit, options.cleanup_pr)
                 i += 1
             else
                 error("--version requires a value")
             end
         elseif arg == "--repo" || arg == "-r"
             if i + 1 <= length(args)
-                options = CLIOptions(options.version, args[i+1], options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit)
+                options = CLIOptions(options.version, args[i+1], options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit, options.audit, options.cleanup_pr)
                 i += 1
             else
                 error("--repo requires a value")
             end
         elseif arg == "--test-commit" || arg == "-t"
             if i + 1 <= length(args)
-                options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, args[i+1])
+                options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, args[i+1], options.audit, options.cleanup_pr)
                 i += 1
             else
                 error("--test-commit requires a commit hash")
             end
         elseif arg == "--dry-run" || arg == "-n"
-            options = CLIOptions(options.version, options.repo, options.help, true, options.validate_branch, options.require_clean, options.test_commit)
+            options = CLIOptions(options.version, options.repo, options.help, true, options.validate_branch, options.require_clean, options.test_commit, options.audit, options.cleanup_pr)
         elseif arg == "--no-validate-branch"
-            options = CLIOptions(options.version, options.repo, options.help, options.dry_run, false, options.require_clean, options.test_commit)
+            options = CLIOptions(options.version, options.repo, options.help, options.dry_run, false, options.require_clean, options.test_commit, options.audit, options.cleanup_pr)
         elseif arg == "--no-require-clean"
-            options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, false, options.test_commit)
+            options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, false, options.test_commit, options.audit, options.cleanup_pr)
+        elseif arg == "--audit" || arg == "-a"
+            options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit, true, options.cleanup_pr)
+        elseif arg == "--cleanup-pr"
+            if i + 1 <= length(args)
+                options = CLIOptions(options.version, options.repo, options.help, options.dry_run, options.validate_branch, options.require_clean, options.test_commit, true, parse(Int, args[i+1]))
+                i += 1
+            else
+                error("--cleanup-pr requires a PR number")
+            end
         else
             error("Unknown argument: $arg")
         end
@@ -120,7 +131,14 @@ function show_help()
     println("  -n, --dry-run           Show what would be done without making changes")
     println("  --no-validate-branch    Skip branch name validation")
     println("  --no-require-clean      Allow dirty working directory")
+    println("  -a, --audit             Run label audit mode (check backport labels)")
+    println("  --cleanup-pr NUMBER     Audit mode: process commits from a specific merged PR")
     println("  -h, --help              Show this help message")
+    println()
+    println("MODES:")
+    println("  Default: Backport PRs with the backport label to the release branch")
+    println("  Audit (--audit): Identify PRs with stale backport labels")
+    println("  Cleanup (--cleanup-pr): Check backport labels for a specific merged PR")
     println()
     println("DEFAULTS:")
     println("  Version is auto-detected from current branch name")
@@ -136,6 +154,9 @@ function show_help()
     println("  julia backporter.jl -r myorg/julia     # Use custom repository")
     println("  julia backporter.jl --dry-run          # Preview changes only")
     println("  julia backporter.jl -t 89dfb68         # Test single commit backport")
+    println("  julia backporter.jl --audit            # Audit backport labels")
+    println("  julia backporter.jl --audit -v 1.11    # Audit specific version")
+    println("  julia backporter.jl --cleanup-pr 1234  # Check labels after PR merge")
 end
 
 # ============================================================================
@@ -708,6 +729,365 @@ function _do_backporting(prs, config::BackportConfig, auth::GitHubAuthenticator)
     end
 end
 
+# ============================================================================
+# Label Audit Functions
+# ============================================================================
+
+struct LabelAuditConfig
+    version::String
+    repo::String
+    github_auth::String
+    backport_label::String
+    release_branch::String
+end
+
+function LabelAuditConfig(version::String, repo::String)
+    github_auth = get(ENV, "GITHUB_TOKEN", "")
+    if isempty(github_auth)
+        error("GITHUB_TOKEN environment variable must be set")
+    end
+    backport_label = "backport $version"
+    release_branch = "release-$version"
+    LabelAuditConfig(version, repo, github_auth, backport_label, release_branch)
+end
+
+function audit_github_headers(auth::String)
+    return Dict(
+        "Authorization" => "token $auth",
+        "Accept" => "application/vnd.github+json",
+        "User-Agent" => "Backporter.jl"
+    )
+end
+
+function find_backport_versions(repo::String, github_auth::String)
+    versions = String[]
+    println("Discovering backport labels...")
+
+    page = 1
+    while true
+        url = "https://api.github.com/repos/$repo/labels?per_page=100&page=$page"
+        response = HTTP.get(url; headers=audit_github_headers(github_auth))
+        data = JSON.parse(String(response.body))
+
+        isempty(data) && break
+
+        for label in data
+            name = label["name"]
+            m = match(r"^backport (\d+\.\d+)$", name)
+            if m !== nothing
+                push!(versions, m.captures[1])
+            end
+        end
+
+        page += 1
+    end
+
+    sort!(versions; by=v -> VersionNumber(v), rev=true)
+    return versions
+end
+
+struct CommitInfo
+    sha::String
+    backport_pr::Union{Int,Nothing}
+end
+
+function extract_pr_numbers_from_message(message::String)
+    prs = Int[]
+    for m in eachmatch(r"\(#(\d+)\)", message)
+        push!(prs, parse(Int, m.captures[1]))
+    end
+    return prs
+end
+
+function extract_merge_pr_from_message(message::String)
+    m = match(r"Merge pull request #(\d+)", message)
+    m !== nothing && return parse(Int, m.captures[1])
+    return nothing
+end
+
+function clone_repo_to_temp(config::LabelAuditConfig)
+    temp_dir = mktempdir()
+    repo_url = "https://github.com/$(config.repo).git"
+    
+    println("Cloning $(config.repo) to temporary directory...")
+    if !success(`git clone --filter=blob:none --no-checkout $repo_url $temp_dir`)
+        rm(temp_dir; force=true, recursive=true)
+        error("Failed to clone repository $(config.repo)")
+    end
+    
+    return temp_dir
+end
+
+function get_commits_from_branch(config::LabelAuditConfig)
+    temp_dir = clone_repo_to_temp(config)
+    
+    try
+        println("Fetching $(config.release_branch)...")
+        if !success(`git -C $temp_dir fetch origin $(config.release_branch)`)
+            error("Failed to fetch $(config.release_branch)")
+        end
+        
+        return parse_git_log_for_commits(temp_dir, "origin/$(config.release_branch)")
+    finally
+        rm(temp_dir; force=true, recursive=true)
+    end
+end
+
+function get_commits_from_pr(config::LabelAuditConfig, pr_number::Int)
+    temp_dir = clone_repo_to_temp(config)
+    
+    try
+        println("Fetching PR #$pr_number...")
+        if !success(`git -C $temp_dir fetch origin pull/$pr_number/head:pr-$pr_number`)
+            error("Failed to fetch PR #$pr_number")
+        end
+        
+        return parse_git_log_for_commits(temp_dir, "pr-$pr_number"; backport_pr=pr_number)
+    finally
+        rm(temp_dir; force=true, recursive=true)
+    end
+end
+
+function parse_git_log_for_commits(repo_dir::String, ref::String; backport_pr::Union{Int,Nothing}=nothing)
+    commits = Dict{Int,CommitInfo}()
+    current_backport_pr = backport_pr
+    
+    println("Parsing git log from $ref...")
+    log_output = read(`git -C $repo_dir log --format=%H%n%B%n---COMMIT_SEPARATOR--- $ref`, String)
+    
+    for commit_block in split(log_output, "---COMMIT_SEPARATOR---")
+        commit_block = strip(commit_block)
+        isempty(commit_block) && continue
+        
+        lines = split(commit_block, '\n')
+        isempty(lines) && continue
+        
+        sha = strip(lines[1])
+        message = join(lines[2:end], '\n')
+        
+        if backport_pr === nothing
+            merge_pr = extract_merge_pr_from_message(message)
+            if merge_pr !== nothing
+                current_backport_pr = merge_pr
+            end
+        end
+        
+        for pr_num in extract_pr_numbers_from_message(message)
+            if !haskey(commits, pr_num)
+                commits[pr_num] = CommitInfo(sha, current_backport_pr)
+            end
+        end
+    end
+    
+    return commits
+end
+
+function get_labeled_closed_prs(config::LabelAuditConfig)
+    prs = []
+
+    println("Fetching closed PRs with label $(config.backport_label)...")
+
+    page = 1
+    while true
+        query = URIs.escapeuri("repo:$(config.repo) is:pr is:closed label:\"$(config.backport_label)\"")
+        url = "https://api.github.com/search/issues?q=$query&per_page=100&page=$page"
+        response = HTTP.get(url; headers=audit_github_headers(config.github_auth))
+        data = JSON.parse(String(response.body))
+
+        items = get(data, "items", [])
+        isempty(items) && break
+
+        for item in items
+            haskey(item, "pull_request") && push!(prs, item)
+        end
+
+        page += 1
+    end
+
+    return prs
+end
+
+function remove_backport_label(config::LabelAuditConfig, pr_number::Int)
+    encoded_label = URIs.escapeuri(config.backport_label)
+    url = "https://api.github.com/repos/$(config.repo)/issues/$pr_number/labels/$encoded_label"
+    HTTP.request("DELETE", url; headers=audit_github_headers(config.github_auth))
+end
+
+struct AuditResult
+    to_remove::Vector{Tuple{Int,String,CommitInfo}}
+    to_keep::Vector{Tuple{Int,String}}
+end
+
+function audit_labels(config::LabelAuditConfig; pr_commits::Union{Dict{Int,CommitInfo},Nothing}=nothing)
+    commits = if pr_commits !== nothing
+        pr_commits
+    else
+        get_commits_from_branch(config)
+    end
+
+    println("Found $(length(commits)) cherry-picked PRs")
+
+    labeled_prs = get_labeled_closed_prs(config)
+    println("Found $(length(labeled_prs)) closed PRs with label $(config.backport_label)")
+
+    to_remove = Tuple{Int,String,CommitInfo}[]
+    to_keep = Tuple{Int,String}[]
+
+    for pr in labeled_prs
+        pr_num = pr["number"]
+        title = pr["title"]
+
+        if haskey(commits, pr_num)
+            push!(to_remove, (pr_num, title, commits[pr_num]))
+        else
+            push!(to_keep, (pr_num, title))
+        end
+    end
+
+    return AuditResult(to_remove, to_keep)
+end
+
+function format_backport_info(info::CommitInfo)
+    bp_str = info.backport_pr !== nothing ? " via #$(info.backport_pr)" : ""
+    return "$(info.sha[1:7])$bp_str"
+end
+
+function print_audit_results(result::AuditResult, config::LabelAuditConfig)
+    println()
+    println("=== PRs already backported (label should be removed) ===")
+    if isempty(result.to_remove)
+        println("None")
+    else
+        for (pr_num, title, info) in result.to_remove
+            println("  #$pr_num: $title ($(format_backport_info(info)))")
+        end
+    end
+
+    println()
+    println("=== PRs still needing backport (label should remain) ===")
+    if isempty(result.to_keep)
+        println("None")
+    else
+        for (pr_num, title) in result.to_keep
+            println("  #$pr_num: $title")
+        end
+    end
+    println()
+end
+
+function apply_audit_changes(result::AuditResult, config::LabelAuditConfig)
+    if isempty(result.to_remove)
+        println("No labels to remove")
+        return
+    end
+
+    println("Removing labels...")
+    for (pr_num, title, info) in result.to_remove
+        try
+            remove_backport_label(config, pr_num)
+            println("  Removed label from #$pr_num")
+        catch e
+            println("  Error processing #$pr_num: $e")
+        end
+    end
+
+    println()
+    println("Done. Removed $(config.backport_label) from $(length(result.to_remove)) PR(s)")
+end
+
+function run_audit_for_version(version::String, repo::String, dry_run::Bool, cleanup_pr::Union{Int,Nothing})
+    if !occursin(r"^\d+\.\d+$", version)
+        error("Invalid version format: $version. Expected X.Y (e.g., 1.13)")
+    end
+
+    config = LabelAuditConfig(version, repo)
+
+    println("Backport Label Audit")
+    println("====================")
+    println("Version: $(config.version)")
+    println("Repository: $(config.repo)")
+    println("Label: $(config.backport_label)")
+    println("Branch: $(config.release_branch)")
+    println("Dry run: $dry_run")
+    println()
+
+    if cleanup_pr !== nothing
+        pr_commits = get_commits_from_pr(config, cleanup_pr)
+        if isempty(pr_commits)
+            println("No cherry-picked PRs found in PR #$cleanup_pr")
+            return
+        end
+        println("Found cherry-picked PRs: $(join(keys(pr_commits), ", "))")
+        result = audit_labels(config; pr_commits=pr_commits)
+        print_audit_results(result, config)
+        if !dry_run
+            apply_audit_changes(result, config)
+        else
+            println("Dry run mode - no changes made")
+            if !isempty(result.to_remove)
+                println("Would remove $(config.backport_label) from $(length(result.to_remove)) PR(s)")
+            end
+        end
+    else
+        result = audit_labels(config)
+        print_audit_results(result, config)
+
+        if dry_run
+            println("Dry run mode - no changes made")
+            if !isempty(result.to_remove)
+                println("Would remove $(config.backport_label) from $(length(result.to_remove)) PR(s)")
+            end
+        else
+            apply_audit_changes(result, config)
+        end
+    end
+end
+
+function run_audit_mode(options::CLIOptions)
+    repo = options.repo
+    if repo === nothing
+        repo = detect_repo_from_remote()
+        if repo === nothing
+            error("Could not detect repository. Please specify with --repo")
+        end
+        println("Auto-detected repository: $repo")
+    end
+
+    if options.version !== nothing
+        run_audit_for_version(options.version, repo, options.dry_run, options.cleanup_pr)
+    else
+        if options.cleanup_pr !== nothing
+            error("--cleanup-pr requires --version to be specified")
+        end
+
+        github_auth = get(ENV, "GITHUB_TOKEN", "")
+        if isempty(github_auth)
+            error("GITHUB_TOKEN environment variable must be set")
+        end
+
+        versions = find_backport_versions(repo, github_auth)
+
+        if isempty(versions)
+            println("No backport labels found in $repo")
+            return
+        end
+
+        println("Found $(length(versions)) backport label(s): $(join(versions, ", "))")
+        println()
+
+        for version in versions
+            run_audit_for_version(version, repo, options.dry_run, nothing)
+            println()
+            println(repeat("=", 60))
+            println()
+        end
+    end
+end
+
+# ============================================================================
+# Main Entry Point
+# ============================================================================
+
 function main(args)
     # Parse command line arguments
     options = parse_cli_args(args)
@@ -717,13 +1097,21 @@ function main(args)
         return
     end
 
-    println("Julia Backporter Tool")
-    println("======================\n")
-
     # Validate environment
     if !ispath(".git")
         error("This script must be run from the root of a git repository")
     end
+
+    # Handle audit mode
+    if options.audit || options.cleanup_pr !== nothing
+        println("Backport Label Audit Tool")
+        println("=========================\n")
+        run_audit_mode(options)
+        return
+    end
+
+    println("Julia Backporter Tool")
+    println("======================\n")
 
     # Validate git state (branch name, clean working directory)
     validate_git_state(options)
